Comment out check_env(), added problem description.
As of now, check_env fails because we do not wrap the angles of the joints to [-pi, pi].  This causes the observations to sometimes not be in the bservation_space. This was a problem before,  but went unnoticed because we always initialized at zero, and the few actions check_env took, never  steered it outside of the observation_space. New issue: how to implement the wrapping?